refactor(test): samle tester i test/unit og test/integration
Flytt unit-tester fra src/ til test/unit/ med katalogstruktur som
speiler src/. Flytt e2e-tester fra e2e/ til test/integration/.
Oppdater vitest.config.ts og playwright.config.ts til nye stier.
Erstatt relative imports i tester med @src/*-alias.
