+++ vitest.config.ts
@@ -13,6 +13,6 @@ export default getViteConfig({
     globals: true,
     environment: "jsdom",
     setupFiles: ["./vitest.setup.ts"],
-    include: ["src/**/*.test.{ts,tsx}"],
+    include: ["test/unit/**/*.test.{ts,tsx}"],
   },
 });
